Extract the StringFormat instance methods into AbstractStringFormat. This makes it easier to create alternative flavors of StringFormat using different placeholder styles such as <placeholder>, %<placeholder>,  etc.
